pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-data-extractor"
version = "1.0.0"
description = "Extract structured data from census and election PDFs"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Wali Reheman", email = "rw8143a@american.edu"}
]
keywords = ["pdf", "ocr", "data-extraction", "census", "election", "tesseract"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing",
    "Topic :: Utilities",
]

dependencies = [
    "requests>=2.31.0",
    "pdf2image>=1.16.3",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "PyYAML>=6.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/wali-reheman/Extract-data-from-old-pdf"
Documentation = "https://github.com/wali-reheman/Extract-data-from-old-pdf#readme"
Repository = "https://github.com/wali-reheman/Extract-data-from-old-pdf"
"Bug Tracker" = "https://github.com/wali-reheman/Extract-data-from-old-pdf/issues"

[project.scripts]
pdf-extract = "pdf_data_extractor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
